Light Blue Con Exp Fixes
Light blue exp was broken in group splits because it was not properly calculating the level difference between the highest group member and the killed NPC for group memembers other than the highest level guy.  This would result in only the highest level guy getting the higher amounts of light blue exp and the lower level players would only get 20%.

Also when calculating raid exp, the con color was being determined on an individual bases instead of using the highest level player in the raid, which was resulting in lower level players getting exp from light blue kills as if they were dark blue.