Added some level checks to summon and enrage
This was done because NPC types have level ranges and sometimes these ranges cross the threshold of summoning and enrage.  Previously the only way to handle this was to duplicate the NPC type and apply the special ability to one type and not the other, but this clutters up the database.  So I hardcoded level checks to fail summon and enrage if the NPC was under the threshold levels even while having the special abilty set.  Currently we have some planar mobs not summoning at all when they should be due to this problem.

I would have checked the maxlevel field but that is not accessible from the Mob class and adding new methods for this would add more clutter so I just checked if the NPC is level 48-50 and 53-55.