Default GCD rule changed to 3500ms; added spell cast times to spell cooldown server checks
Sony's global cooldown was 3.5 seconds and this is very obvious from logs.  I changed the rule default value to 3500ms, so the default is now accuruate and non-custom.  However the database has the rule in it so that will need to be changed for Quarm.

Spell cast times were added to the server cooldowns for individual spells.  This will prevent chain casts of certain spells with a cooldown, like enchanter nukes and wizard pillars.  Before, the cooldown was run concurrently with the spell cast time when the spell was in gem slot 1 and clickables were used, which in some cases could nearly double the cast rate.  I subtracted 1 second from the cast time added to the cooldown just to be cautious.