update code and database.cpp code: * reduced cone of vision angle when checking if an NPC can see a sneaking player * fix LFG inconsistently showing after zoning. * fixed an issue that was causing spell dealy on priority 0 spell list. * #reload Command Overhaul. * added more utility to devtools. * added zone store (import from eqemu) * convert race, gender, and class to constexpr. * Race constants refactor. * Clean up classes. * clean up ShowQuickStats * Fix Luclin AC Softcap Calculation Typo (credit to nickgal) * added virtual zonepoints and #showzonepoints command.cpp * fixed corpse exploits (credit to PQ) * fixed an issue where client_max_level (databucket, qglobal, quest api) not allowing leveling to end of level * readd hotzone rule.
Torven commits and explanation:
_______________________________
Archery and Throw attacks now unparryable
Old logs and Live server tests confirm that these attacks should not be parryable.
Note that they SHOULD be blockable, and my leaving them blockable is intended.

_________________________________________________________________________________

Improved SE_HitChance, SE_IncreaseArchery and SE_AvoidMeleeChance (includes SQL)

These spell effects are supposed to multiply to-hit and avoidance AC, not be independent rolls.  This commit should theoretically make the results precise or nearly so.

How I know this is the case:

A player did archery hit tests in Jan 2002, shooting over 12,000 arrows on Ssolet Dnaas, and calculated the miss rates with and without Eagle Eye.  The results matched a to-hit multiplier of 1.4 and not an independent success roll of 40%.

Melee discipline descriptions say "Increase Chance to Avoid Melee by 10000%" and "Increase Chance to Hit by 10000% with All Skills" which makes no sense unless it's a to-hit or avoidance AC multiplier.

I found a Steel Warrior post of a warrior mentioning that hits could very rarely go through Fortitude disc.  Multiplying avoidance AC by 10,000% results in a hit rate of roughly 1 in 150 against the player.  Sony's own discipline description also says "increasing your chance" and not "all".  An independent roll the way we previously did it would prevent all hits.

Furthermore the math simply fits well doing it this way, is very simple and is the most obvious way to do it so that's a strong sign it's correct.

NOTE: This commit requires changes to the spell data for Evasive and Precision, which Cavedude put in as custom spells years ago and made up values for.  He made up values for these discs because he didn't know how the logic should work and it was before my melee revamp which computed accurate combat stats.  Had he used the values listed on Allakhazam/Lucy it would have resulted in the disciplines being far too good using the old logic.  SQL is included with this commit.  This makes the values match Lucy's, although Precision had its pentalty removed when discs were made into spells so that's why it's not shown in Lucy.  I left Trueshot's 33% in because I don't know what to set it but since it doesn't stack with Eagle Eye (logs suggest that is correct) it's not much of an issue if it's wrong.

The ArcheryHitPenalty rule was also removed.  The reason why it was there is because the independent roll was making Eagle Eye way too good, so this rule cut that buff in half to make the hit rate reasonable.  The fixed logic removed the need for it.

With this new logic/math and fixed discipline values and archery rule removal, the resulting hit rates will end up similar to before and not really nerf or buff anything very much although Evasive will be a little bit worse, Fortitude + Voiddance will not be 100% immunity and Archery with Eagle Eye will hit about 2% better.

______________________________________________________________

NPCs made to not flee if top hater has less HP
See this: https://web.archive.org/web/20030706022528/http://necro.eqclasses.com/forum/post.asp?method=ReplyQuote&REPLY_ID=174331&TOPIC_ID=19007&FORUM_ID=2

Other posts also casually mention it, like this: https://web.archive.org/web/20050129102418/http://necro.eqclasses.com/viewtopic.php?p=11773&amp